Implemented a controlable car.
The car class has been implemented and it can be controlled with w/s to avance/retrocede and a/d to turn.
A file that has a class aimed to implement a state machine to manage keyboard inputs has been created.
A file with minor math functions has been created.